refactor: move strategies inside folder and add jwt package
