+++ src/auth/auth.service.ts
@@ -1,6 +1,9 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { User, UserService } from '../user/user.service';
 
+
+
+export type UserInfo = Omit<User, "password"> 
 @Injectable()
 export class AuthService {
 
@@ -8,12 +11,13 @@ export class AuthService {
     constructor(private readonly userService: UserService){}
 
      
-    validateUser(email:string, password:string) : User {
+    validateUser(email:string, pass:string) : UserInfo{
         const user =  this.userService.getUserByUserEmail(email)
-        if(!user || user.password !== password){
+        if(!user || user.password !== pass){
             throw new UnauthorizedException()
         }
-        return user;
+        const {passowrd, ...result} = user;
+        return result;
     }
 
 }
